add user

+++ pages/api/domains/[id].ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { deleteDomainById } from '../../../models/Domain';
+import { deleteDomainById, getDomainById } from '../../../models/Domain';
 import dbConnect from '../../../lib/dbConnect';
 
 (async function () {
@@ -14,6 +14,16 @@ export default async function handleDomains(
   if (req.method === 'DELETE') {
     try {
       const id = req.query.id as string;
+      const userId = req.headers.userid;
+      const domain = await getDomainById(id);
+      if (!domain) {
+        res.status(404).json({ message: 'Invalid domain name' });
+        return;
+      }
+      if (domain.userId !== userId) {
+        res.status(403).json({ message: 'Permission denied!' });
+        return;
+      }
       await deleteDomainById(id);
       res.status(204).json({ id });
     } catch (error) {
